Add the job dispatcher
The job dispatcher's purpose is to send jobs coming from Disque to workers
that can process the job, interpret the worker's return code and handle
failed jobs.

To do its job, the job dispatcher uses these classes:

- JobRouter knows what jobs go to which workers
- FailureStrategyRepository knows how a failed job should be handled

JobRouter in turn uses these classes:
- WorkerDirections are directions to a worker (its URL, class name, command)
- Route knows the right directions for a set of queues
- Call knows how exactly a worker should be called and how the result should
be interpreted
- CallFactory turns WorkerDirections into Calls, which are returned
to the JobDispatcher

FailureStrategyRepository works as follows:
- FailureStrategy is an exact description of what should happen to a failed job
- The repository contains all known FailureStrategies - they are read from
the service container and can also be added manually during bootstrap
- If a job fails, the job dispatcher asks the repository for a failure strategy
according to the queue to which the job belongs
- The repository looks into the configuration what strategy the user has
assigned to the queue and returns it
- If no strategy or an unknown failure strategy has been configured,
the repository returns a default strategy

Successful jobs are simply removed from Disque (ACKed).