Include a retry count test in the NACKing of a job
When NACKing a job, we check if it has lifetime left before its next retry.
I have added another test that checks if it has still allowed retries left.

This commit also adds a new public function to log the failed job. It is needed
in various places when handling a failed job, so it is useful to have it here.