Remove the code coverage icon from the README
To explain for anyone who is interested: The coverage shows a low number, 35%.
This is caused not by a lack of tests, but by the fact that a big part
of Disqontrol is a command-line application:
Defining parameters and options in console commands, spawning new processes and
cleaning them up, managing interrupt signals etc. This is all code that cannot
be unit-tested. The rest is tested by unit as well as by integration tests,
as you can see in the directory `tests/`.